// kd3619-roktools-main/kd-frontend/types.ts
export type PlayerHonorHistory = {
  id: string;
  name: string;
  history: {
    fileName: string;
    fileId: string;
    date?: string;
    honorPoint: number;
  }[];
};